Move riddle test
