+++ composer.json
@@ -125,7 +125,7 @@
         "drupal/pathauto": "^1.0",
         "drupal/responsive_preview": "^1.0",
         "drupal/redirect": "^1.0",
-        "drupal/riddle_marketplace": "dev-feature/sort-riddles",
+        "drupal/riddle_marketplace": "dev-2.x",
         "drupal/scheduler": "^1.0",
         "drupal/select2": "^1.0",
         "drupal/simple_sitemap": "^2.0",
